[libSyntax] Push creation of deferred nodes to the ParsedRawSyntaxRecorder
The ParsedRawSyntaxRecorder is already responsible for creating recorded
nodes, now it is also responsible for creating deferred nodes. This
has two advantages:
1. All the ParsedRawSyntaxNode creation logic is in one place
2. We can push the creation of deferred nodes further down to the
   SyntaxParseActions, which might be able to create deferred nodes more
   efficiently.
